Restructure to separate Nuwave\Lighthouse\Bind namespace
